file structure, autloader, refactor db func to db class, replace config with const
