+++ app/config.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App;
+
+class config {
+	const DBNAME = 'chat_app';
+	const HOST = '127.0.0.1:3306';
+	const USER = 'root';
+	const PASSWORD = 'root';
+}
+
