+++ app/helpers/Tools.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Helpers;
+
+class Tools
+{
+	public static function get_ip()
+	{
+		$value = '';
+		if (!empty($_SERVER['HTTP_CLIENT_IP'])) {
+			$value = $_SERVER['HTTP_CLIENT_IP'];
+		} elseif (!empty($_SERVER['HTTP_X_FORWARDED_FOR'])) {
+			$value = $_SERVER['HTTP_X_FORWARDED_FOR'];
+		} elseif (!empty($_SERVER['REMOTE_ADDR'])) {
+			$value = $_SERVER['REMOTE_ADDR'];
+		}
+
+		return $value;
+	}
+}
+
+
