+++ app/views/partials/footer.php
