--- database.sql
@@ -1,67 +1,0 @@
--- MySQL Script generated by MySQL Workbench
--- Thu Jan  9 12:36:42 2025
--- Model: New Model    Version: 1.0
--- MySQL Workbench Forward Engineering
-
-SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0;
-SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0;
-SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='ONLY_FULL_GROUP_BY,STRICT_TRANS_TABLES,NO_ZERO_IN_DATE,NO_ZERO_DATE,ERROR_FOR_DIVISION_BY_ZERO,NO_ENGINE_SUBSTITUTION';
-
--- -----------------------------------------------------
--- Schema mydb
--- -----------------------------------------------------
--- -----------------------------------------------------
--- Schema chat_app
--- -----------------------------------------------------
-
--- -----------------------------------------------------
--- Schema chat_app
--- -----------------------------------------------------
-CREATE SCHEMA IF NOT EXISTS `chat_app` DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_0900_ai_ci ;
-USE `chat_app` ;
-
--- -----------------------------------------------------
--- Table `chat_app`.`user`
--- -----------------------------------------------------
-CREATE TABLE IF NOT EXISTS `chat_app`.`user` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `username` VARCHAR(45) NOT NULL,
-  `email` VARCHAR(255) NOT NULL,
-  `passwordHash` VARCHAR(255) NOT NULL,
-  `create_date` TIMESTAMP NULL DEFAULT CURRENT_TIMESTAMP,
-  `register_ip` VARCHAR(45) NOT NULL,
-  `browser` TEXT NULL DEFAULT NULL,
-  PRIMARY KEY (`id`))
-ENGINE = InnoDB
-AUTO_INCREMENT = 22
-DEFAULT CHARACTER SET = utf8mb4
-COLLATE = utf8mb4_0900_ai_ci;
-
-
--- -----------------------------------------------------
--- Table `chat_app`.`messages`
--- -----------------------------------------------------
-CREATE TABLE IF NOT EXISTS `chat_app`.`messages` (
-  `id` INT NOT NULL AUTO_INCREMENT,
-  `text` TEXT NOT NULL,
-  `filePath` VARCHAR(255) NULL DEFAULT NULL,
-  `create_date` TIMESTAMP NULL DEFAULT CURRENT_TIMESTAMP,
-  `user_id` INT NOT NULL,
-  `sender_ip` VARCHAR(45) NOT NULL,
-  `browser` TEXT NULL DEFAULT NULL,
-  PRIMARY KEY (`id`),
-  INDEX `user_id_idx` (`user_id` ASC) VISIBLE,
-  CONSTRAINT `user_id`
-    FOREIGN KEY (`user_id`)
-    REFERENCES `chat_app`.`user` (`id`)
-    ON DELETE CASCADE
-    ON UPDATE CASCADE)
-ENGINE = InnoDB
-AUTO_INCREMENT = 155
-DEFAULT CHARACTER SET = utf8mb4
-COLLATE = utf8mb4_0900_ai_ci;
-
-
-SET SQL_MODE=@OLD_SQL_MODE;
-SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS;
-SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS;
